requirements.txt
django
django-compressor
django-filter
django-money
djangorestframework
gunicorn
ipython
markdown
matplotlib
pillow
psycopg2-binary
py-moneyed
pypdf2
python-dateutil
python-resize-image
requests
sendgrid
stripe
weasyprint
